+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## [Unreleased](https://github.com/Goooler/golang-gradle-plugin/compare/0.2.2...HEAD) - 2026-xx-xx
+
 
 ## [0.2.2](https://github.com/Goooler/golang-gradle-plugin/releases/tag/0.2.2) - 2026-05-07