+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 
-- Prioritize PATH over GOROOT for Go executable resolution.
+- Prioritize `PATH` over `GOROOT` for Go executable resolution.
 
 ## [0.2.2](https://github.com/Goooler/golang-gradle-plugin/releases/tag/0.2.2) - 2026-05-07